// biomeSchema.d.ts
export type Schema = string
// ...
export type Glob = string
export type ImportGroups = ImportGroup[]
export type RuleAssistConfigurationFor_Null = RuleAssistPlainConfiguration | RuleAssistWithOptionsFor_Null
export type Bool = boolean
export type IndentStyle = 'tab' | 'space'
export type IndentWidth = number
export type LineEnding = 'lf' | 'crlf' | 'cr'
/**
 * Validated value for the `line_width` formatter options
 *
 * The allowed range of values is 1..=320
 */
export type LineWidth = number
export type QuoteStyle = 'double' | 'single'
export type MaxSize = number
export type AttributePosition = 'auto' | 'multiline'
/**
 * Put the `>` of a multi-line HTML or JSX element at the end of the last line instead of being alone on the next line (does not apply to self closing elements).
 */
export type BracketSameLine = boolean
export type BracketSpacing = boolean
export type Expand = 'auto' | 'always' | 'never'
// ...
export type GroupPlainConfiguration = 'off' | 'on' | 'info' | 'warn' | 'error'
// ...
export type SeverityOrGroupFor_Correctness = GroupPlainConfiguration | Correctness
// ...
export type Overrides = OverridePattern[]
export type PluginConfiguration = string
export type Plugins = PluginConfiguration[]
export type VcsClientKind = 'git'

/**
 * The configuration that is contained inside the file `biome.json`
 */
export interface Configuration {
  /**
   * A field for the [JSON schema](https://json-schema.org/) specification
   */
  $schema?: Schema | null
  /**
   * Specific configuration for assists
   */
  assist?: AssistConfiguration | null
  /**
   * Specific configuration for the Css language
   */
  css?: CssConfiguration | null
  /**
   * A list of paths to other JSON files, used to extends the current configuration.
   */
  extends?: string[] | null
  /**
   * The configuration of the filesystem
   */
  files?: FilesConfiguration | null
  /**
   * The configuration of the formatter
   */
  formatter?: FormatterConfiguration | null
  /**
   * Specific configuration for the GraphQL language
   */
  graphql?: GraphqlConfiguration | null
  /**
   * Specific configuration for the GraphQL language
   */
  grit?: GritConfiguration | null
  /**
   * Specific configuration for the HTML language
   */
  html?: HtmlConfiguration | null
  /**
   * Specific configuration for the JavaScript language
   */
  javascript?: JsConfiguration | null
  /**
   * Specific configuration for the Json language
   */
  json?: JsonConfiguration | null
  /**
   * The configuration for the linter
   */
  linter?: LinterConfiguration | null
  /**
   * A list of granular patterns that should be applied only to a sub set of files
   */
  overrides?: Overrides | null
  /**
   * List of plugins to load.
   */
  plugins?: Plugins | null
  /**
   * Indicates whether this configuration file is at the root of a Biome project. By default, this is `true`.
   */
  root?: Bool | null
  /**
   * The configuration of the VCS integration
   */
  vcs?: VcsConfiguration | null
}
export interface AssistConfiguration {
  /**
   * Whether Biome should fail in CLI if the assist were not applied to the code.
   */
  actions?: Actions | null
  /**
   * Whether Biome should enable assist via LSP.
   */
  enabled?: Bool | null
  /**
   * A list of glob patterns. Biome will include files/folders that will match these patterns.
   */
  includes?: Glob[] | null
}
export interface Actions {
  /**
   * It enables the assist actions recommended by Biome. `true` by default.
   */
  recommended?: boolean | null
  source?: Source | null
}
// ...
export interface CssConfiguration {
  /**
   * CSS assist options
   */
  assist?: CssAssistConfiguration | null
  /**
   * CSS formatter options
   */
  formatter?: CssFormatterConfiguration | null
  /**
   * CSS globals
   */
  globals?: string[] | null
  /**
   * CSS linter options
   */
  linter?: CssLinterConfiguration | null
  /**
   * CSS parsing options
   */
  parser?: CssParserConfiguration | null
}
/**
 * Options that changes how the CSS assist behaves
 */
export interface CssAssistConfiguration {
  /**
   * Control the assist for CSS files.
   */
  enabled?: Bool | null
}
/**
 * Options that changes how the CSS formatter behaves
 */
export interface CssFormatterConfiguration {
  /**
   * Control the formatter for CSS (and its super languages) files.
   */
  enabled?: Bool | null
  /**
   * The indent style applied to CSS (and its super languages) files.
   */
  indentStyle?: IndentStyle | null
  /**
   * The size of the indentation applied to CSS (and its super languages) files. Default to 2.
   */
  indentWidth?: IndentWidth | null
  /**
   * The type of line ending applied to CSS (and its super languages) files.
   */
  lineEnding?: LineEnding | null
  /**
   * What's the max width of a line applied to CSS (and its super languages) files. Defaults to 80.
   */
  lineWidth?: LineWidth | null
  /**
   * The type of quotes used in CSS code. Defaults to double.
   */
  quoteStyle?: QuoteStyle | null
}
/**
 * Options that changes how the CSS linter behaves
 */
export interface CssLinterConfiguration {
  /**
   * Control the linter for CSS files.
   */
  enabled?: Bool | null
}
/**
 * Options that changes how the CSS parser behaves
 */
export interface CssParserConfiguration {
  /**
   * Allow comments to appear on incorrect lines in `.css` files
   */
  allowWrongLineComments?: Bool | null
  /**
   * Enables parsing of CSS Modules specific features.
   */
  cssModules?: Bool | null
}
/**
 * The configuration of the filesystem
 */
export interface FilesConfiguration {
  /**
   * Tells Biome to not emit diagnostics when handling files that doesn't know
   */
  ignoreUnknown?: Bool | null
  /**
   * A list of glob patterns. Biome will handle only those files/folders that will match these patterns.
   */
  includes?: Glob[] | null
  /**
   * The maximum allowed size for source code files in bytes. Files above this limit will be ignored for performance reasons. Defaults to 1 MiB
   */
  maxSize?: MaxSize | null
}
/**
 * Generic options applied to all files
 */
export interface FormatterConfiguration {
  /**
   * The attribute position style in HTML-ish languages. Defaults to auto.
   */
  attributePosition?: AttributePosition | null
  /**
   * Put the `>` of a multi-line HTML or JSX element at the end of the last line instead of being alone on the next line (does not apply to self closing elements).
   */
  bracketSameLine?: BracketSameLine | null
  /**
   * Whether to insert spaces around brackets in object literals. Defaults to true.
   */
  bracketSpacing?: BracketSpacing | null
  enabled?: Bool | null
  /**
   * Whether to expand arrays and objects on multiple lines. When set to `auto`, object literals are formatted on multiple lines if the first property has a newline, and array literals are formatted on a single line if it fits in the line. When set to `always`, these literals are formatted on multiple lines, regardless of length of the list. When set to `never`, these literals are formatted on a single line if it fits in the line. When formatting `package.json`, Biome will use `always` unless configured otherwise. Defaults to "auto".
   */
  expand?: Expand | null
  /**
   * Stores whether formatting should be allowed to proceed if a given file has syntax errors
   */
  formatWithErrors?: Bool | null
  /**
   * A list of glob patterns. The formatter will include files/folders that will match these patterns.
   */
  includes?: Glob[] | null
  /**
   * The indent style.
   */
  indentStyle?: IndentStyle | null
  /**
   * The size of the indentation, 2 by default
   */
  indentWidth?: IndentWidth | null
  /**
   * The type of line ending.
   */
  lineEnding?: LineEnding | null
  /**
   * What's the max width of a line. Defaults to 80.
   */
  lineWidth?: LineWidth | null
  /**
   * Use any `.editorconfig` files to configure the formatter. Configuration in `biome.json` will override `.editorconfig` configuration.
   *
   * Default: `true`.
   */
  useEditorconfig?: Bool | null
}
/**
 * Options applied to GraphQL files
 */
export interface GraphqlConfiguration {
  /**
   * Assist options
   */
  assist?: GraphqlAssistConfiguration | null
  /**
   * GraphQL formatter options
   */
  formatter?: GraphqlFormatterConfiguration | null
  linter?: GraphqlLinterConfiguration | null
}
/**
 * Options that changes how the GraphQL linter behaves
 */
export interface GraphqlAssistConfiguration {
  /**
   * Control the formatter for GraphQL files.
   */
  enabled?: Bool | null
}
/**
 * Options that changes how the GraphQL formatter behaves
 */
export interface GraphqlFormatterConfiguration {
  /**
   * Whether to insert spaces around brackets in object literals. Defaults to true.
   */
  bracketSpacing?: BracketSpacing | null
  /**
   * Control the formatter for GraphQL files.
   */
  enabled?: Bool | null
  /**
   * The indent style applied to GraphQL files.
   */
  indentStyle?: IndentStyle | null
  /**
   * The size of the indentation applied to GraphQL files. Default to 2.
   */
  indentWidth?: IndentWidth | null
  /**
   * The type of line ending applied to GraphQL files.
   */
  lineEnding?: LineEnding | null
  /**
   * What's the max width of a line applied to GraphQL files. Defaults to 80.
   */
  lineWidth?: LineWidth | null
  /**
   * The type of quotes used in GraphQL code. Defaults to double.
   */
  quoteStyle?: QuoteStyle | null
}
/**
 * Options that change how the GraphQL linter behaves.
 */
export interface GraphqlLinterConfiguration {
  /**
   * Control the formatter for GraphQL files.
   */
  enabled?: Bool | null
}
/**
 * Options applied to GritQL files
 */
export interface GritConfiguration {
  /**
   * Assist options
   */
  assist?: GritAssistConfiguration | null
  /**
   * Formatting options
   */
  formatter?: GritFormatterConfiguration | null
  /**
   * Formatting options
   */
  linter?: GritLinterConfiguration | null
}
export interface GritAssistConfiguration {
  /**
   * Control the assist functionality for Grit files.
   */
  enabled?: Bool | null
}
export interface GritFormatterConfiguration {
  /**
   * Control the formatter for Grit files.
   */
  enabled?: Bool | null
  /**
   * The indent style applied to Grit files.
   */
  indentStyle?: IndentStyle | null
  /**
   * The size of the indentation applied to Grit files. Default to 2.
   */
  indentWidth?: IndentWidth | null
  /**
   * The type of line ending applied to Grit files.
   */
  lineEnding?: LineEnding | null
  /**
   * What's the max width of a line applied to Grit files. Defaults to 80.
   */
  lineWidth?: LineWidth | null
}
export interface GritLinterConfiguration {
  /**
   * Control the linter for Grit files.
   */
  enabled?: Bool | null
}
/**
 * Options applied to HTML files
 */
export interface HtmlConfiguration {
  /**
   * HTML formatter options
   */
  formatter?: HtmlFormatterConfiguration | null
  /**
   * HTML parsing options
   */
  parser?: HtmlParserConfiguration
}
// ...
export interface JsConfiguration {
  /**
   * Assist options
   */
  assist?: JsAssistConfiguration | null
  /**
   * Formatting options
   */
  formatter?: JsFormatterConfiguration | null
  /**
   * A list of global bindings that should be ignored by the analyzers
   *
   * If defined here, they should not emit diagnostics.
   */
  globals?: string[] | null
  /**
   * Indicates the type of runtime or transformation used for interpreting JSX.
   */
  jsxRuntime?: JsxRuntime | null
  /**
   * Linter options
   */
  linter?: JsLinterConfiguration | null
  /**
   * Parsing options
   */
  parser?: JsParserConfiguration | null
}
// ...
export interface JsonConfiguration {
  /**
   * Assist options
   */
  assist?: JsonAssistConfiguration | null
  /**
   * Formatting options
   */
  formatter?: JsonFormatterConfiguration | null
  /**
   * Linting options
   */
  linter?: JsonLinterConfiguration | null
  /**
   * Parsing options
   */
  parser?: JsonParserConfiguration | null
}
// ...
export interface LinterConfiguration {
  /**
   * An object where the keys are the names of the domains, and the values are `all`, `recommended`, or `none`.
   */
  domains?: RuleDomains | null
  /**
   * if `false`, it disables the feature and the linter won't be executed. `true` by default
   */
  enabled?: Bool | null
  /**
   * A list of glob patterns. The analyzer will handle only those files/folders that will match these patterns.
   */
  includes?: Glob[] | null
  /**
   * List of rules
   */
  rules?: Rules | null
}
// ...
export interface VcsConfiguration {
  /**
   * The kind of client.
   */
  clientKind?: VcsClientKind | null
  /**
   * The main branch of the project
   */
  defaultBranch?: string | null
  /**
   * Whether Biome should integrate itself with the VCS client
   */
  enabled?: Bool | null
  /**
   * The folder where Biome should check for VCS files. By default, Biome will use the same folder where `biome.json` was found.
   *
   * If Biome can't find the configuration, it will attempt to use the current working directory. If no current working directory can't be found, Biome won't use the VCS integration, and a diagnostic will be emitted
   */
  root?: string | null
  /**
   * Whether Biome should use the VCS ignore file. When [true], Biome will ignore the files specified in the ignore file.
   */
  useIgnoreFile?: Bool | null
}